jormungandr: corrections
